+++ mitra_models/src/custom_feeds/types.rs
@@ -1,0 +1,10 @@
+use postgres_types::FromSql;
+use uuid::Uuid;
+
+#[derive(FromSql)]
+#[postgres(name = "custom_feed")]
+pub struct DbCustomFeed {
+    pub id: i32,
+    pub owner_id: Uuid,
+    pub feed_name: String,
+}